+++ src/com/intellij/idea/plugin/hybris/project/descriptors/impl/YRegularModuleDescriptor.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -62,6 +62,12 @@ abstract class YRegularModuleDescriptor protected constructor(
 
         requiredExtensionNames.addAll(getAdditionalRequiredExtensionNames())
 
+        if (hasWebModule) {
+            requiredExtensionNames
+                .map { "$it." + HybrisConstants.COMMON_WEB_MODULE_DIRECTORY }
+                .filter { moduleDescriptors.contains(it) }
+                .let { requiredExtensionNames.addAll(it) }
+        }
         if (hasHmcModule) {
             requiredExtensionNames.add(HybrisConstants.EXTENSION_NAME_HMC)
         }