+++ src/com/intellij/idea/plugin/hybris/project/factories/SubModuleDescriptorFactory.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -30,15 +30,10 @@ object SubModuleDescriptorFactory {
     fun buildAll(owner: YRegularModuleDescriptor): Set<YSubModuleDescriptor> {
         val subModules = mutableSetOf<YSubModuleDescriptor>()
 
-        if (owner.hasWebModule) {
-            build(owner, HybrisConstants.WEB_MODULE_DIRECTORY, subModules) { YWebSubModuleDescriptor(owner, it) }
-        }
-        if (owner.hasHmcModule) {
-            build(owner, HybrisConstants.HMC_MODULE_DIRECTORY, subModules) { YHmcSubModuleDescriptor(owner, it) }
-        }
-        if (owner.isHacAddon) {
-            build(owner, HybrisConstants.HAC_MODULE_DIRECTORY, subModules) { YHacSubModuleDescriptor(owner, it) }
-        }
+        if (owner.hasWebModule) build(owner, HybrisConstants.WEB_MODULE_DIRECTORY, subModules) { YWebSubModuleDescriptor(owner, it) }
+        if (owner.hasHmcModule) build(owner, HybrisConstants.HMC_MODULE_DIRECTORY, subModules) { YHmcSubModuleDescriptor(owner, it) }
+        if (owner.isHacAddon) build(owner, HybrisConstants.HAC_MODULE_DIRECTORY, subModules) { YHacSubModuleDescriptor(owner, it) }
+
         if (owner.hasBackofficeModule) {
             build(owner, HybrisConstants.BACKOFFICE_MODULE_DIRECTORY, subModules) { backoffice ->
                 val subModule = YBackofficeSubModuleDescriptor(owner, backoffice)
@@ -51,6 +46,7 @@ object SubModuleDescriptorFactory {
                 subModule
             }
         }
+
         build(owner, HybrisConstants.COMMON_WEB_MODULE_DIRECTORY, subModules) { YCommonWebSubModuleDescriptor(owner, it) }
         build(owner, HybrisConstants.ACCELERATOR_ADDON_WEB_PATH, subModules) { YAcceleratorAddonSubModuleDescriptor(owner, it) }