+++ CHANGELOG.md
@@ -22,6 +22,9 @@
 - Allow modification of the non-project root files [#794](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/794)
 - Adjust type system merge for `Product` type (it extends both `GenericItem` and `LocalizableItem`) [#785](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/785)
 
+### Fixes
+- OOTB `commonweb` sub-modules are not re-imported on project refresh [#795](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/795)
+
 ## [2023.2.10]
 
 ### `Project Import` enhancements